requirements_base.txt
Flask
Flask-Admin
Flask-Migrate
flask-restful
Flask-Script
Flask-SQLAlchemy
Flask-WTF

psycopg2
